Unset Credit card payment in auction if product's price is over Pay Now limit
We can't list products with price over Pay Now limit if Pay Now option
is enabled TradeMe account. And we can't disable Pay Now for such products
when we list them. So we can only remove Credit card payment this there
will be no Pay Now option on the auction